+++ multilint/__init__.py
@@ -34,7 +34,7 @@ except ImportError:
 
 __author__ = 'Adam Johnson'
 __email__ = 'me@adamj.eu'
-__version__ = '2.0.0'
+__version__ = '2.0.1'
 
 
 def main():